创建 msgmapper.go
